extensions to ServiceInstaller
