--- Api/BccPay.Core.Infrastructure/Extensions/RefitExtensions.cs
@@ -1,18 +1,0 @@
-﻿using BccPay.Core.Infrastructure.PaymentProviders.RefitClients;
-using Microsoft.Extensions.DependencyInjection;
-using Refit;
-using System;
-
-namespace BccPay.Core.Infrastructure.Payments.Extensions
-{
-    public static class RefitExtensions
-    {
-        public static IServiceCollection AddRefitClients(this IServiceCollection services)
-        {
-            services.AddRefitClient<INetsClient>()
-                .ConfigureHttpClient(client => client.BaseAddress = new Uri("https://test.api.dibspayment.eu"));
-
-            return services;
-        }
-    }
-}
