+++ Api/BccPay.Core.Infrastructure/PaymentProviders/IPaymentProviderFactory.cs
@@ -2,6 +2,6 @@
 {
     public interface IPaymentProviderFactory
     {
-        IPaymentProvider GetPaymentProvider(string paymentMethod);
+        IPaymentProvider GetPaymentProvider(string paymentMethod); // NOTE: maybe enumeration instead of string?
     }
 }
